# pkg/trace/writer: switch from flushing based on spans to size