--- a/pkg/trace/config/config_test.go
+++ b/pkg/trace/config/config_test.go
@@ -224,9 +224,8 @@ func TestFullIniConfig(t *testing.T) {
 	}, c.StatsWriterConfig)
 
 	assert.Equal(writerconfig.TraceWriterConfig{
-		MaxSpansPerPayload: 100,
-		FlushPeriod:        3 * time.Second,
-		UpdateInfoPeriod:   2 * time.Second,
+		FlushPeriod:      3 * time.Second,
+		UpdateInfoPeriod: 2 * time.Second,
 		SenderConfig: writerconfig.QueuablePayloadSenderConf{
 			MaxAge:            time.Second,
 			MaxQueuedBytes:    456,
@@ -350,7 +349,6 @@ func TestUndocumentedYamlConfig(t *testing.T) {
 	assert.Equal(50, c.MaxConnections)
 
 	// Assert Trace Writer
-	assert.Equal(11, c.TraceWriterConfig.MaxSpansPerPayload)
 	assert.Equal(22*time.Second, c.TraceWriterConfig.FlushPeriod)
 	assert.Equal(33*time.Second, c.TraceWriterConfig.UpdateInfoPeriod)
 	assert.Equal(15*time.Second, c.TraceWriterConfig.SenderConfig.MaxAge)