--- a/pkg/trace/config/testdata/full.ini
+++ b/pkg/trace/config/testdata/full.ini
@@ -73,7 +73,6 @@ exp_backoff_base_milliseconds: 1
 exp_backoff_growth_base: 2
 
 [trace.writer.traces]
-max_spans_per_payload: 100
 flush_period_seconds: 3
 update_info_period_seconds: 2
 queue_max_age_seconds: 1