--- a/pkg/trace/writer/trace_nix_test.go
+++ b/pkg/trace/writer/trace_nix_test.go
@@ -3,13 +3,14 @@
 package writer
 
 import (
+	"errors"
 	"fmt"
-	"math"
 	"strings"
 	"testing"
 	"time"
 
 	"github.com/DataDog/datadog-agent/pkg/trace/info"
+	"github.com/DataDog/datadog-agent/pkg/trace/test/testutil"
 	"github.com/stretchr/testify/assert"
 )
 
@@ -51,141 +52,121 @@ func TestPeriodicFlush(t *testing.T) {
 func TestPeriodicStats(t *testing.T) {
 	assert := assert.New(t)
 
+	defer func(old int) {
+		payloadFlushThreshold = old // reset original setting
+	}(payloadFlushThreshold)
+
 	testFlushPeriod := 100 * time.Millisecond
 
 	// Create a trace writer, its incoming channel and the endpoint that receives the payloads
 	traceWriter, traceChannel, testEndpoint, statsClient := testTraceWriter()
-	traceWriter.conf.FlushPeriod = 100 * time.Millisecond
+	traceWriter.conf.FlushPeriod = testFlushPeriod
 	traceWriter.conf.UpdateInfoPeriod = 100 * time.Millisecond
-	traceWriter.conf.MaxSpansPerPayload = 10
 	traceWriter.Start()
 
-	var (
-		expectedNumPayloads       int64
-		expectedNumSpans          int64
-		expectedNumTraces         int64
-		expectedNumBytes          int64
-		expectedNumErrors         int64
-		expectedMinNumRetries     int64
-		expectedNumSingleMaxSpans int64
-	)
+	want := make(map[string]int) // maps metric name to expected sum
 
-	// Send a bunch of sampled traces that should go together in a single payload
-	payload1SampledTraces := []*TracePackage{
-		randomTracePackage(2, 0),
-		randomTracePackage(2, 0),
-		randomTracePackage(2, 0),
-	}
-	expectedNumPayloads++
-	expectedNumSpans += 6
-	expectedNumTraces += 3
-	expectedNumBytes += calculateTracePayloadSize(payload1SampledTraces)
-
-	for _, sampledTrace := range payload1SampledTraces {
-		traceChannel <- sampledTrace
+	payload1 := []*TracePackage{
+		randomTracePackage(10, 0),
+		randomTracePackage(10, 0),
+		randomTracePackage(10, 0),
 	}
 
-	// Send a single trace that goes over the span limit
-	payload2SampledTraces := []*TracePackage{
-		randomTracePackage(20, 0),
-	}
-	expectedNumPayloads++
-	expectedNumSpans += 20
-	expectedNumTraces++
-	expectedNumBytes += calculateTracePayloadSize(payload2SampledTraces)
-	expectedNumSingleMaxSpans++
+	// this payload will fit
+	payloadFlushThreshold = int(calculateTracePayloadEstimatedSize(payload1))
 
-	for _, sampledTrace := range payload2SampledTraces {
-		traceChannel <- sampledTrace
+	for _, trace := range payload1 {
+		traceChannel <- trace
 	}
 
-	// Wait for twice the flush period
-	time.Sleep(2 * testFlushPeriod)
+	want["spans"] = 30
+	want["traces"] = len(payload1)
+	want["bytes_estimated"] = calculateTracePayloadEstimatedSize(payload1)
+	want["bytes_uncompressed"] = calculateTracePayloadSize(payload1)
 
-	// Send a third payload with other 3 traces with an errored out endpoint
-	testEndpoint.SetError(fmt.Errorf("non retriable error"))
-	payload3SampledTraces := []*TracePackage{
-		randomTracePackage(2, 0),
-		randomTracePackage(2, 0),
-		randomTracePackage(2, 0),
+	pkg1 := randomTracePackage(90, 0)
+	traceChannel <- pkg1 // pushes us over AND flushes itself due to huge size
+
+	want["payloads"] = 2
+	want["traces"]++
+	want["spans"] += 90
+	want["bytes_estimated"] += calculateTracePayloadEstimatedSize([]*TracePackage{pkg1})
+	want["bytes_uncompressed"] += calculateTracePayloadSize([]*TracePackage{pkg1})
+	want["single_max_size"]++
+
+	// this will fit again
+	for _, trace := range payload1 {
+		traceChannel <- trace
 	}
 
-	expectedNumErrors++
-	expectedNumTraces += 3
-	expectedNumSpans += 6
-	expectedNumBytes += calculateTracePayloadSize(payload3SampledTraces)
+	time.Sleep(2 * testFlushPeriod) // flush
 
-	for _, sampledTrace := range payload3SampledTraces {
-		traceChannel <- sampledTrace
+	want["payloads"]++
+	want["traces"] += len(payload1)
+	want["spans"] += 30 // spans in payload1
+	want["bytes_estimated"] += calculateTracePayloadEstimatedSize(payload1)
+	want["bytes_uncompressed"] += calculateTracePayloadSize(payload1)
+
+	// trigger some errors
+	testEndpoint.SetError(errors.New("non retriable error"))
+
+	for _, trace := range payload1 {
+		traceChannel <- trace
 	}
 
-	// Wait for twice the flush period
-	time.Sleep(2 * testFlushPeriod)
+	time.Sleep(2 * testFlushPeriod) // flush
 
-	// And then send a fourth payload with other 3 traces with an errored out endpoint but retriable
+	want["errors"] = 1
+	want["traces"] += len(payload1)
+	want["spans"] += 30 // spans in payload1
+	want["bytes_estimated"] += calculateTracePayloadEstimatedSize(payload1)
+	want["bytes_uncompressed"] += calculateTracePayloadSize(payload1)
+
+	// trigger retriable error
 	testEndpoint.SetError(&retriableError{
-		err:      fmt.Errorf("non retriable error"),
+		err:      errors.New("retriable error"),
 		endpoint: testEndpoint,
 	})
-	payload4SampledTraces := []*TracePackage{
+
+	payload4 := []*TracePackage{
 		randomTracePackage(2, 0),
 		randomTracePackage(2, 0),
 		randomTracePackage(2, 0),
 	}
 
-	expectedMinNumRetries++
-	expectedNumTraces += 3
-	expectedNumSpans += 6
-	expectedNumBytes += calculateTracePayloadSize(payload4SampledTraces)
-
-	for _, sampledTrace := range payload4SampledTraces {
+	for _, sampledTrace := range payload4 {
 		traceChannel <- sampledTrace
 	}
 
-	// Wait for twice the flush period to see at least one retry
 	time.Sleep(2 * testFlushPeriod)
+
+	want["retries"] = 2
+	want["traces"] += len(payload1)
+	want["spans"] += 6
+	want["bytes_estimated"] += calculateTracePayloadEstimatedSize(payload4)
+	want["bytes_uncompressed"] += calculateTracePayloadSize(payload4)
 
 	// Close and stop
 	close(traceChannel)
 	traceWriter.Stop()
 
-	// Then we expect some counts to have been sent to the stats client for each update tick (there should have been
-	// at least 3 ticks)
-	countSummaries := statsClient.GetCountSummaries()
+	expectStats := summaryCompareFunc(assert, statsClient)
+	for key, sum := range want {
+		delta := 1.
+		if key == "bytes_uncompressed" {
+			delta = 1000.
+		}
+		expectStats("datadog.trace_agent.trace_writer."+key, 7, sum, delta)
+	}
+}
 
-	// Payload counts
-	payloadSummary := countSummaries["datadog.trace_agent.trace_writer.payloads"]
-	assert.True(len(payloadSummary.Calls) >= 3, "There should have been multiple payload count calls")
-	assert.Equal(expectedNumPayloads, payloadSummary.Sum)
-
-	// Traces counts
-	tracesSummary := countSummaries["datadog.trace_agent.trace_writer.traces"]
-	assert.True(len(tracesSummary.Calls) >= 3, "There should have been multiple traces count calls")
-	assert.Equal(expectedNumTraces, tracesSummary.Sum)
-
-	// Spans counts
-	spansSummary := countSummaries["datadog.trace_agent.trace_writer.spans"]
-	assert.True(len(spansSummary.Calls) >= 3, "There should have been multiple spans count calls")
-	assert.Equal(expectedNumSpans, spansSummary.Sum)
-
-	// Bytes counts
-	bytesSummary := countSummaries["datadog.trace_agent.trace_writer.bytes"]
-	assert.True(len(bytesSummary.Calls) >= 3, "There should have been multiple bytes count calls")
-	// FIXME: Is GZIP non-deterministic? Why won't equal work here?
-	assert.True(math.Abs(float64(expectedNumBytes-bytesSummary.Sum)) < 100., "Bytes should be within expectations")
-
-	// Retry counts
-	retriesSummary := countSummaries["datadog.trace_agent.trace_writer.retries"]
-	assert.True(len(retriesSummary.Calls) >= 2, "There should have been multiple retries count calls")
-	assert.True(retriesSummary.Sum >= expectedMinNumRetries)
-
-	// Error counts
-	errorsSummary := countSummaries["datadog.trace_agent.trace_writer.errors"]
-	assert.True(len(errorsSummary.Calls) >= 3, "There should have been multiple errors count calls")
-	assert.Equal(expectedNumErrors, errorsSummary.Sum)
-
-	// Single trace max spans
-	singleMaxSpansSummary := countSummaries["datadog.trace_agent.trace_writer.single_max_spans"]
-	assert.True(len(singleMaxSpansSummary.Calls) >= 3, "There should have been multiple single max spans count calls")
-	assert.Equal(expectedNumSingleMaxSpans, singleMaxSpansSummary.Sum)
+func summaryCompareFunc(assert *assert.Assertions, statsClient *testutil.TestStatsClient) func(name string, calls int, sum int, delta float64) {
+	summaries := statsClient.GetCountSummaries()
+	return func(name string, minCalls int, sum int, delta float64) {
+		summary, ok := summaries[name]
+		assert.True(ok, fmt.Sprintf("%s not found", name))
+		calls := len(summary.Calls)
+		assert.True(calls >= minCalls, fmt.Sprintf("%s: expected at least %d calls, got %d", name, minCalls, calls))
+		assert.InDelta(sum, summary.Sum, delta, fmt.Sprintf("%s: expected %d, got %d", name, sum, summary.Sum))
+	}
 }