This change modifies the flushing logic to flush based on an estimated
payload size as opposed to a maximum number of spans.

Adding the payload size computation slightly decreases the
trace-handling method's speed, but is negligible considering that the
only thing this function is doing is technically two `append` calls.

```
benchmark                         old ns/op     new ns/op     delta
BenchmarkHandleSampledTrace-4     1117          1373          +22.92%
```